requirements.txt
python-telegram-bot==20.6
pdfplumber==0.7.6
python-dotenv==1.0.0
openai==1.0.0   # optional: only if you want auto-generated questions with OpenAI